Trying out sending messages back and forward
View plugins are configured to listen for messages from the model, and
also the widgets defined for that plugin are configured to send messages
to the model. Model plugins can then registed to listen for messages.

In this case, changing the subject line triggers an update message for
the key 'subject', which the MIMELite model plugin is registered to
process (which it then does nothing with).

Likewise, if the MIMELite object has its subject changed, it will send
and update message to the view, which the Compose view plugin is
configured to receive. That sub is not defined, so the program will
crash.

The general idea being that any view plugin can listen for any model
message and via-versa. Ie: New plugins could listen for subject line
changes and perform some action on that.