Onchange events kind of supported
I wanted to have the subject in the list of composed mails be updated
realtime when you typed out the subject. To do this I created a special
type of SENDS key called onchange. The message is not sent to the model
but is instead just fired. In this case it triggers the Compose plugin
to fetch the NewMailList listbox and update the value.

I think perhaps this shouldn't be a magic SENDS key, because that should
pertain to model messages only. Maybe I will make a ONCHANGE key in the
messages hash that will install separate onchange events.